Move `getErrorOptions`
